don't turn off curve padding in the result report
For now, don't turn off padding in the plots of the
result HTML report.

Turning off padding does not work for the main curve
for whatever reason. The result is that main curve and
error curve are displayed in different time scales.

This confuses and annoyes users, bette generate curve with
same time scale.